Create Efficient Frontier Theorem
The efficient frontier is the set of optimal portfolios that offer the highest expected return for a defined level of risk or the lowest risk for a given level of expected return. Portfolios that lie below the efficient frontier are sub-optimal because they do not provide enough return for the level of risk.

This code can be used to add stocks/ assets to receive an optimal portfolio allocation for various assets. Can be used for asset allocation over a range of assets and further optimise the investment allocations within each asset class.